Request: The 64-bit Quark

+++ include/SymbolClass.h
@@ -140,7 +140,7 @@ namespace MFM{
 
     void printBitSizeOfClass();
 
-    bool getDefaultQuark(u32& dqref);
+    bool getDefaultQuark(u64& dqref);
     bool getPackedDefaultValue(u64& dpkref);
     bool getDefaultValue(BV8K& dvref); //return true if ready
